Add logic to reduce colours
This logic reduces the colours in the image so that each grid
cell has the same colour across **all** of its pixels.
Unfortunately, this meant that the dimensions of the original
image have to be altered and the image has to be cropped so as
to adjust to the new grid.